pyproject.toml
[tool.poetry]
name = "poker-night-bot"
version = "0.1.0"
description = ""
authors = ["Uri Maimon <uri.maimon@medtronic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.0.0"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
